Saving Dinos (Herbivores)
I got the save system to work with saving and loading the locations, rotation, and scale of the herbivores in the world. I'm only gonna focus on the herbivores for now because I want to have this all smoothed out so that I can essentially copy and paste everything to work with the carnivores. The issue is that this will work as long as the dinos that were saved are already present/spawned in the world when loading. This is not very ideal, so I'm going to try and make it so that I can spawn in the dinos without making numerous duplicates every time a game is loaded.