Saving Herbivores (UPDATE)
I seemed to have fixed the previous issue. I was having an issue where when loading, however many herbivores were saved would just be spawned in addition to the number of herbivores already present on the map (it was amusing to watch, but it's definitely not what you want). So, I fixed this by getting all actors of the herbivore class and destroying them all before getting the save and spawning them in.